schemaValidation: rename .js → .mjs for unambiguous ESM + mock-api null-filter
volto-hydra's package.json declares "type": "commonjs", so .js files
imported from TS/Vite test runners get treated as CommonJS — breaks
named imports of the pure validation functions. Renaming to .mjs
forces ESM regardless of package.json. The pure file has no Volto
dependencies so this doesn't propagate.

schemaInheritance.js now imports from ./schemaValidation.mjs.

Also: mock-plone-api.cjs's @search endpoint was throwing TypeError
when iterating content dirs that don't have a parseable data.json
(e.g. screenshot images uploaded via Volto). Filter null content
before formatSearchItem rather than letting it throw.
