+++ packages/volto-hydra/src/utils/schemaValidation.mjs
